Moved parmyay into its own package. Added code coverage generation
Updated build scripts to account for this.
