+++ api/cover.out
@@ -1,0 +1,124 @@
+mode: set
+parmyay\achievements.go:15.38,18.36 3 1
+parmyay\achievements.go:18.36,20.3 1 1
+parmyay\achievements.go:20.3,20.42 1 1
+parmyay\achievements.go:20.42,23.3 2 1
+parmyay\achievements.go:23.3,25.3 1 1
+parmyay\achievements.go:29.38,33.2 3 1
+parmyay\achievements.go:36.37,41.25 4 1
+parmyay\achievements.go:41.25,43.3 1 1
+parmyay\achievements.go:43.3,45.3 1 1
+parmyay\achievements.go:49.40,54.25 4 1
+parmyay\achievements.go:54.25,57.39 3 1
+parmyay\achievements.go:57.39,64.4 3 1
+parmyay\achievements.go:64.4,66.4 1 1
+parmyay\achievements.go:68.3,70.3 1 1
+parmyay\achievements.go:74.40,80.25 4 1
+parmyay\achievements.go:80.25,83.3 2 1
+parmyay\achievements.go:83.3,85.3 1 1
+parmyay\achievements.go:89.41,93.2 3 0
+parmyay\main.go:19.27,20.43 1 1
+parmyay\main.go:23.2,23.14 1 1
+parmyay\main.go:20.43,22.3 1 1
+parmyay\main.go:27.29,28.30 1 1
+parmyay\main.go:28.30,31.3 2 1
+parmyay\main.go:35.15,38.70 2 0
+parmyay\main.go:43.2,43.34 1 0
+parmyay\main.go:46.2,46.29 1 0
+parmyay\main.go:49.2,49.27 1 0
+parmyay\main.go:53.2,53.28 1 0
+parmyay\main.go:38.70,39.77 1 0
+parmyay\main.go:40.3,40.19 1 0
+parmyay\main.go:43.34,45.3 1 0
+parmyay\main.go:46.29,48.3 1 0
+parmyay\main.go:49.27,52.3 2 0
+parmyay\main.go:53.28,55.3 1 0
+parmyay\main.go:65.51,68.2 2 0
+parmyay\main.go:71.50,72.15 1 1
+parmyay\main.go:75.2,75.21 1 1
+parmyay\main.go:72.15,74.3 1 1
+parmyay\main.go:79.54,80.21 1 1
+parmyay\main.go:83.2,84.18 2 1
+parmyay\main.go:87.2,92.2 4 1
+parmyay\main.go:102.2,103.2 2 1
+parmyay\main.go:113.2,114.2 2 1
+parmyay\main.go:124.2,125.2 2 1
+parmyay\main.go:134.2,134.15 1 1
+parmyay\main.go:80.21,82.3 1 1
+parmyay\main.go:84.18,86.3 1 1
+parmyay\main.go:92.2,100.3 7 1
+parmyay\main.go:103.2,111.3 7 1
+parmyay\main.go:114.2,122.3 7 1
+parmyay\main.go:125.2,133.3 7 1
+parmyay\reviews.go:24.33,28.91 3 0
+parmyay\reviews.go:28.91,34.19 5 0
+parmyay\reviews.go:38.3,40.20 3 0
+parmyay\reviews.go:44.3,47.40 4 0
+parmyay\reviews.go:34.19,37.4 2 0
+parmyay\reviews.go:40.20,43.4 2 0
+parmyay\reviews.go:48.3,50.3 1 0
+parmyay\reviews.go:54.33,58.2 3 1
+parmyay\reviews.go:61.32,66.20 4 1
+parmyay\reviews.go:66.20,68.3 1 1
+parmyay\reviews.go:68.3,70.3 1 1
+parmyay\reviews.go:74.35,79.46 4 0
+parmyay\reviews.go:79.46,80.21 1 0
+parmyay\reviews.go:80.21,93.4 5 0
+parmyay\reviews.go:93.4,95.4 1 0
+parmyay\reviews.go:97.3,99.3 1 0
+parmyay\reviews.go:103.35,108.20 4 1
+parmyay\reviews.go:108.20,119.3 5 1
+parmyay\reviews.go:119.3,121.3 1 1
+parmyay\reviews.go:125.36,129.2 3 0
+parmyay\test.go:12.72,18.16 5 1
+parmyay\test.go:21.2,23.35 3 1
+parmyay\test.go:18.16,20.3 1 0
+parmyay\test.go:23.35,25.3 1 0
+parmyay\test.go:29.58,33.16 3 1
+parmyay\test.go:36.2,38.35 3 1
+parmyay\test.go:33.16,35.3 1 0
+parmyay\test.go:38.35,40.3 1 0
+parmyay\test.go:44.61,48.16 3 1
+parmyay\test.go:51.2,53.35 3 1
+parmyay\test.go:48.16,50.3 1 0
+parmyay\test.go:53.35,55.3 1 0
+parmyay\test.go:59.65,63.16 3 1
+parmyay\test.go:66.2,68.35 3 1
+parmyay\test.go:63.16,65.3 1 0
+parmyay\test.go:68.35,70.3 1 0
+parmyay\test.go:74.71,80.16 5 1
+parmyay\test.go:83.2,85.35 3 1
+parmyay\test.go:80.16,82.3 1 0
+parmyay\test.go:85.35,87.3 1 0
+parmyay\users.go:26.31,29.116 3 0
+parmyay\users.go:29.116,34.3 4 0
+parmyay\users.go:34.3,36.3 1 0
+parmyay\users.go:40.31,45.2 3 1
+parmyay\users.go:48.30,53.18 4 1
+parmyay\users.go:53.18,55.3 1 1
+parmyay\users.go:55.3,57.3 1 1
+parmyay\users.go:61.33,67.44 4 0
+parmyay\users.go:67.44,68.19 1 0
+parmyay\users.go:68.19,84.4 5 0
+parmyay\users.go:84.4,86.4 1 0
+parmyay\users.go:88.3,90.3 1 0
+parmyay\users.go:94.33,99.18 4 1
+parmyay\users.go:99.18,106.3 5 1
+parmyay\users.go:106.3,108.3 1 1
+parmyay\users.go:112.34,116.2 3 0
+parmyay\venues.go:26.32,30.113 3 0
+parmyay\venues.go:30.113,35.3 4 0
+parmyay\venues.go:35.3,37.3 1 0
+parmyay\venues.go:41.32,45.2 3 1
+parmyay\venues.go:48.31,53.19 4 1
+parmyay\venues.go:53.19,55.3 1 1
+parmyay\venues.go:55.3,57.3 1 1
+parmyay\venues.go:61.34,66.113 4 0
+parmyay\venues.go:66.113,67.20 1 0
+parmyay\venues.go:67.20,82.4 5 0
+parmyay\venues.go:82.4,84.4 1 0
+parmyay\venues.go:86.3,88.3 1 0
+parmyay\venues.go:92.34,97.19 4 1
+parmyay\venues.go:97.19,108.3 5 1
+parmyay\venues.go:108.3,110.3 1 1
+parmyay\venues.go:114.35,118.2 3 0
