+++ api/src/parmyay/achievements.go
@@ -1,4 +1,4 @@
-package main
+package parmyay
 
 import (
 	"github.com/gin-gonic/gin"
