+++ api/src/parmyay/reviews.go
@@ -1,4 +1,4 @@
-package main
+package parmyay
 
 import (
 	"strconv"
