+++ api/src/parmyay/reviews_test.go
@@ -1,4 +1,4 @@
-package main
+package parmyay
 
 import (
 	"testing"
