+++ api/src/parmyay/users_test.go
@@ -1,4 +1,4 @@
-package main
+package parmyay
 
 import (
 	"testing"
