+++ api/src/parmyay/venues.go
@@ -1,4 +1,4 @@
-package main
+package parmyay
 
 import (
 	"time"
