+++ api/src/parmyay/venues_test.go
@@ -1,4 +1,4 @@
-package main
+package parmyay
 
 import (
 	"testing"
